Update README.md
I used Chat GPT to try to understand this problem. I tried talking with the TA about it as well and also looked at other students repositories. Going through other repositories seems to not help at all as I am not sure what it is I am suppose to be doing. I Think I may be starting to get the idea of what it is we are suppose to be doing. But I still really feel like I have no  idea what is happening or why. I am hoping that the more I go through the assignments I will start to make sense.  I feel like I need to be walked through a multitude of examples very slowly till I start to get the idea.